requirements.txt
# Python 3.10+ is recommended
sqlalchemy>=2.0.4

# Core dependencies
phonenumbers>=8.12.0
email-validator>=1.1.0
python-dateutil>=2.8.0

# NLP dependencies for hybrid NLP system
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99  # For some tokenizers
accelerate>=0.20.0  # For faster model loading
scikit-learn>=1.0.0  # For training scripts

# Hugging Face Hub for downloading/uploading models
huggingface_hub>=0.16.0  # For model download/upload